print status after each command
